FILE-290: Having separate routes file for the Admin
