+++ app/uk/gov/hmrc/fileupload/controllers/FileUploadController.scala
@@ -20,14 +20,11 @@ import cats.data.Xor
 import play.api.libs.iteratee.Enumerator
 import play.api.libs.json.{JsObject, JsString, Json}
 import play.api.mvc._
-import reactivemongo.api.commands.WriteResult
 import uk.gov.hmrc.fileupload.controllers.FileUploadController._
 import uk.gov.hmrc.fileupload.fileupload._
 import uk.gov.hmrc.fileupload.notifier.NotifierService._
 import uk.gov.hmrc.fileupload.quarantine.FileInQuarantineStored
-import uk.gov.hmrc.fileupload.transfer.TransferRequested
 import uk.gov.hmrc.fileupload.utils.errorAsJson
-import uk.gov.hmrc.fileupload.virusscan.VirusScanRequested
 import uk.gov.hmrc.fileupload.{EnvelopeId, FileId, FileRefId}
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -63,16 +60,6 @@ class FileUploadController(uploadParser: () => BodyParser[MultipartFormData[Futu
           Future.successful(BadRequest(errorAsJson("Request must have exactly 1 file attached")))
         }
     }
-  }
-
-  def scan(envelopeId: EnvelopeId, fileId: FileId, fileRefId: FileRefId) = Action.async { request =>
-    notify(VirusScanRequested(envelopeId = envelopeId, fileId = fileId, fileRefId = fileRefId))
-    Future.successful(Ok)
-  }
-
-  def transfer(envelopeId: EnvelopeId, fileId: FileId, fileRefId: FileRefId) = Action.async { request =>
-    notify(TransferRequested(envelopeId = envelopeId, fileId = fileId, fileRefId = fileRefId))
-    Future.successful(Ok)
   }
 
 }
