+++ app/uk/gov/hmrc/fileupload/frontendGlobal.scala
@@ -25,7 +25,7 @@ import play.api.mvc.Request
 import play.api.{Mode => _, _}
 import play.twirl.api.Html
 import uk.gov.hmrc.crypto.ApplicationCrypto
-import uk.gov.hmrc.fileupload.controllers.{FileUploadController, UploadParser}
+import uk.gov.hmrc.fileupload.controllers.{AdminAppController, FileUploadController, UploadParser}
 import uk.gov.hmrc.fileupload.infrastructure.{DefaultMongoConnection, HttpStreamingBody, PlayHttp}
 import uk.gov.hmrc.fileupload.notifier.NotifierService.NotifyResult
 import uk.gov.hmrc.fileupload.notifier.{NotifierRepository, NotifierService}
@@ -74,6 +74,7 @@ object FrontendGlobal
     Akka.system.actorOf(TransferActor.props(subscribe, streamTransferCall), "transferActor")
 
     fileUploadController
+    AdminAppController
     testOnlyController
   }
 
@@ -142,8 +143,10 @@ object FrontendGlobal
   lazy val sendNotification = NotifierRepository.send(auditedHttpExecute, ServiceConfig.fileUploadBackendBaseUrl) _
 
   lazy val fileUploadController = new FileUploadController(uploadParser = uploadParser, notify = notifyAndPublish, now = now)
+  lazy val AdminAppController = new AdminAppController(uploadParser = uploadParser, notify = notifyAndPublish, now = now)
 
   private val FileUploadControllerClass = classOf[FileUploadController]
+  private val AdminAppControllerClass = classOf[AdminAppController]
 
   lazy val testOnlyController = new TestOnlyController(ServiceConfig.fileUploadBackendBaseUrl)
   private val TestOnlyControllerClass = classOf[TestOnlyController]
@@ -151,6 +154,7 @@ object FrontendGlobal
   override def getControllerInstance[A](controllerClass: Class[A]): A = {
     controllerClass match {
       case FileUploadControllerClass => fileUploadController.asInstanceOf[A]
+      case AdminAppControllerClass => AdminAppController.asInstanceOf[A]
       case TestOnlyControllerClass => testOnlyController.asInstanceOf[A]
       case _ => super.getControllerInstance(controllerClass)
     }
