+++ conf/prod.routes
@@ -2,3 +2,4 @@
 GET        /admin/metrics        com.kenshoo.play.metrics.MetricsController.metrics
 ->         /                     health.Routes
 ->         /                     app.Routes
+->         /admin                admin.Routes
